Implement JT_ENUM
JT_ENUM creates a list strings of the enum. This makes it possible to
create a TypeHandler for the enum and recognizing the strings. This does
not work if the enums are assigned a value, and there exists an assert to
catch this.

JT_ENUM_DECLARE_STRING_PARSER can be used to generate this TypeHandler
specialisation. This declaration can not be used inside namespaces!!! Also
if the enum is declared inside a namespace then use
JT_ENUM_NAMESPACE_DECLARE_STRING_PARSER.